feedback 2
Remove typing from requirements
Move put methods to a roundtrip test class. Fix super() so this works in roundtrip tests with multiple inheritance.
