+++ metadata_service/proxy/atlas_proxy.py
@@ -8,7 +8,7 @@ from typing import Any, Dict, List, Union, Optional
 
 from amundsen_common.models.dashboard import DashboardSummary
 from amundsen_common.models.popular_table import PopularTable
-from amundsen_common.models.table import Application, Column, Stat, Table, Tag, User, Reader, \
+from amundsen_common.models.table import Column, Stat, Table, Tag, User, Reader, \
     ProgrammaticDescription, ResourceReport
 from amundsen_common.models.user import User as UserEntity
 from atlasclient.client import Atlas
@@ -997,34 +997,3 @@ class AtlasProxy(BaseProxy):
                                   id: str,
                                   resource_type: ResourceType) -> Dict[str, List[DashboardSummary]]:
         return {}
-
-    def put_user(self, *, data: User) -> None:
-        pass
-
-    def post_users(self, *, data: List[User]) -> None:
-        pass
-
-    def put_app(self, *, data: Application) -> None:
-        pass
-
-    def post_apps(self, *, data: List[Application]) -> None:
-        pass
-
-    def put_table(self, *, table: Table) -> None:
-        pass
-
-    def post_tables(self, *, tables: List[Table]) -> None:
-        pass
-
-    def put_column(self, *, table_uri: str, column: Column) -> None:
-        pass
-
-    def put_programmatic_table_description(self, *, table_uri: str, description: ProgrammaticDescription) -> None:
-        pass
-
-    def put_programmatic_column_description(self, *, table_uri: str, column_name: str,
-                                            description: Description) -> None:
-        pass
-
-    def add_read_count(self, *, table_uri: str, user_id: str, read_count: int) -> None:
-        pass
