+++ requirements.txt
@@ -31,12 +31,11 @@ pytest-mock==1.1
 # PEP 484
 # License: PSF
 # Upstream url: https://github.com/python/typing
-typing==3.6.4
 typing-extensions==3.7.4
 
 # A common package that holds the models deifnition and schemas that are used
 # accross different amundsen repositories.
-amundsen-common==0.5.2
+amundsen-common==0.5.3
 amundsen-gremlin==0.0.4
 
 boto3==1.12.12
