+++ setup.py
@@ -23,6 +23,7 @@ setup(
     include_package_data=True,
     zip_safe=False,
     dependency_links=[],
+    install_requires=requirements,
     extras_require={
         'oidc': ['flaskoidc==0.0.2']
     },
