+++ tests/unit/proxy/roundtrip/test_janus_graph_proxy.py
@@ -4,13 +4,12 @@
 from typing import Any, Mapping
 import unittest
 
-from metadata_service.proxy.janus_graph_proxy import JanusGraphGremlinProxy
-
 from .abstract_gremlin_proxy_tests import abstract_gremlin_proxy_test_class
+from .roundtrip_janusgraph_proxy import RoundtripJanusGraphProxy
 
 
 class JanusGraphGremlinProxyTest(
         abstract_gremlin_proxy_test_class(), unittest.TestCase):  # type: ignore
-    def _create_gremlin_proxy(self, config: Mapping[str, Any]) -> JanusGraphGremlinProxy:
+    def _create_gremlin_proxy(self, config: Mapping[str, Any]) -> RoundtripJanusGraphProxy:
         # Don't use PROXY_HOST, PROXY_PORT, PROXY_PASSWORD.  They might not be JanusGraph
-        return JanusGraphGremlinProxy(host=config['JANUS_GRAPH_URL'])
+        return RoundtripJanusGraphProxy(host=config['JANUS_GRAPH_URL'])
