+++ tests/unit/proxy/roundtrip/test_neptune_proxy.py
@@ -13,14 +13,15 @@ from gremlin_python.process.traversal import Cardinality
 from gremlin_python.process.graph_traversal import __
 
 from .abstract_gremlin_proxy_tests import abstract_gremlin_proxy_test_class
+from .roundtrip_neptune_proxy import RoundtripNeptuneGremlinProxy
 
 
 class NeptuneGremlinProxyTest(
         abstract_gremlin_proxy_test_class(), unittest.TestCase):  # type: ignore
-    def _create_gremlin_proxy(self, config: Mapping[str, Any]) -> NeptuneGremlinProxy:
+    def _create_gremlin_proxy(self, config: Mapping[str, Any]) -> RoundtripNeptuneGremlinProxy:
         # Don't use PROXY_HOST, PROXY_PORT, PROXY_PASSWORD.  They might not be neptune
-        return NeptuneGremlinProxy(host=config['NEPTUNE_URL'], password=config['NEPTUNE_SESSION'],
-                                   neptune_bulk_loader_s3_bucket_name=config['NEPTUNE_BULK_LOADER_S3_BUCKET_NAME'])
+        return RoundtripNeptuneGremlinProxy(host=config['NEPTUNE_URL'], password=config['NEPTUNE_SESSION'],
+                                            neptune_bulk_loader_s3_bucket_name=config['NEPTUNE_BULK_LOADER_S3_BUCKET_NAME']) # noqa E501
 
     def test_is_retryable(self) -> None:
         exception = gremlin_python.driver.protocol.GremlinServerError(dict(
